Fix failing XArray tests
The issue was in calling `cat` on AxisArrays and KeyedArrays.

For AxisArrays the issue was it wouldn't allow two arrays to have the
same axis values.

KeyedArrays allows cat-ing two arrays with the same values but the default
axis keys generated in the tests did not match those from the code.